Remove all traces of the in memory shopping list.
This severs almost all ties between the app and RtM. The only exception
is the export feature; other than that, the shopping list is purely
tracked in the database.
